changes to match data with firebase

+++ app/src/main/java/ict2105/team02/application/ui/main/QRScannerActivity.kt
@@ -90,7 +90,7 @@ class QRScannerActivity : AppCompatActivity() {
                     if (serial.isNotEmpty()) {
                         runOnUiThread {
                             cameraSource!!.stop()
-                            val fragment = ScopeDetailFragment.newInstance(serial)
+                            val fragment = ScopeDetailFragment.newInstance(serial.toInt())
                             fragment.show(supportFragmentManager, "scope_detail")
                             // showToast(scannedValue)
                         }
